+++ vignette-currency-converter/vignette-currency-converter.php
@@ -3,7 +3,7 @@
  * Plugin Name: Vignette Currency Converter
  * Plugin URI: https://github.com/Moley123/WooCommerceMultiCurrency
  * Description: Multi-currency support for vignette products with ExchangeRate-API integration
- * Version: 1.3.0
+ * Version: 1.4.0
  * Author: Mark Lebrett
  * Author URI: https://marklebrett.co.uk
  * License: GPL v2 or later
@@ -21,7 +21,7 @@ if (!defined('ABSPATH')) {
 }
 
 // Define plugin constants
-define('VCC_VERSION', '1.3.0');
+define('VCC_VERSION', '1.4.0');
 define('VCC_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('VCC_PLUGIN_URL', plugin_dir_url(__FILE__));
 define('VCC_PLUGIN_BASENAME', plugin_basename(__FILE__));